// libs/workspace/src/builders/packager/schema.ts
/**
 * @license
 * Copyright 2022 Dynatrace LLC
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface PackagerAssetDef {
  glob: string;
  input: string;
  output: string;
}

/** The options that can be used with the packager builder */
export interface PackagerOptions {
  buildTarget: string;
  releasePackageJson: string;
  versionPlaceholder: string;
  ngVersionPlaceholder: string;
  styles: PackagerAssetDef[];
  assets: PackagerAssetDef[];
  additionalTargets: string[];
}